requirements.txt
pytest
pytest-cov
instagram-private-api
random-poem
Pillow
instabot
instagrapi
bs4
lxml
pywebview